Fix: se agregaron funciones que estaban en script.js a eliminarControlador
